make use of set VERSION_TAG
Use case:
In pam-personbruker we have to separate applications stemming from
the same codebase. Since changes in the code directly affect both
applications, we would like to deploy both applications simultaneously.
Taking that into consideration, it would be better if both packages for
the applications contain the same version tag.